Epsilon riduzione; Prodotto di partizioni (esplicito); Verificato l'utilizzo ottimale di memoria del programma e stima di questo.

+++ ising_simulation.h
@@ -49,8 +49,8 @@ public:
     int *border3;
     int border_size;
     const config_t *config_reference() { return config;}
-    int energia_cinetica();
-    int energia_magnetica();
+    double energia_cinetica();
+    double energia_magnetica();
     double magnetizzazione();
 };
